requirements.txt
scipy==1.4.1
tensorflow==1.13.1
h5py==2.9.0
matplotlib==3.1.1
Keras==2.2.4
numpy==1.17.3
streamlit==0.52.2
scikit_image==0.15.0
imgaug==0.3.0
ipython==7.12.0
Pillow==7.0.0
